debug pose iphone

+++ scripts/splatam_paralell.py
@@ -178,13 +178,19 @@ def run_splatam(args):
     seq_experiment = SourceFileLoader(
             os.path.basename(config_file), config_file
         ).load_module()
-    
+
+    aniso = 'aniso' if experiment_args['aniso'] else 'iso'
+    if experiment_args['aniso']:
+        seq_experiment.config['gaussian_distribution'] = "anisotropic"
+    else:
+        seq_experiment.config['gaussian_distribution'] = "isotropic"
+
     if 'jono' in seq_experiment.config['data']['gradslam_data_cfg']:
         # seq = os.path.join(seq, 'ims/27')
         tracking_iters_cam = 0
-        run_name = f"splatam_{seq}_{experiment_args['seed']}_{experiment_args['mov_init_by']}_{experiment_args['tracking_iters']}_{experiment_args['tracking_iters_init']}_{experiment_args['tracking_iters_cam']}_{experiment_args['num_frames']}_{experiment_args['feature_dim']}_{experiment_args['init_jono']}_{experiment_args['jono_depth']}_{experiment_args['remove_gaussians']}_{experiment_args['sil_thres_gaussians']}_{experiment_args['l1_losses_embedding']}_{experiment_args['l1_losses_color']}_{experiment_args['bg_reg']}_{experiment_args['embeddings_lr']}_{experiment_args['red_lr']}_{experiment_args['red_lr_cam']}_{experiment_args['embedding_weight']}_{experiment_args['use_seg_for_nn']}_{experiment_args['weight_iso']}_{experiment_args['exp_weight']}_{experiment_args['loss_weight_emb']}_{experiment_args['loss_weight_iso']}_{experiment_args['loss_weight_rigid']}_{experiment_args['loss_weight_rot']}_{experiment_args['loss_weight_depth_cam']}_{experiment_args['forward_propagate_camera']}_{experiment_args['trafo_mat']}_{experiment_args['feats_224']}_{experiment_args['restart_if_fail']}_{experiment_args['early_stop']}_{experiment_args['stride']}_{experiment_args['time_window']}_{experiment_args['l1_losses_scale']}_{experiment_args['last_x']}_{experiment_args['kNN']}_{experiment_args['desired_image_height']}_{experiment_args['desired_image_width']}_{experiment_args['instseg_obj']}_{experiment_args['instseg_cam']}_transformed"
+        run_name = f"splatam_{seq}_{experiment_args['seed']}_{experiment_args['mov_init_by']}_{experiment_args['tracking_iters']}_{experiment_args['tracking_iters_init']}_{experiment_args['tracking_iters_cam']}_{experiment_args['num_frames']}_{experiment_args['feature_dim']}_{experiment_args['init_jono']}_{experiment_args['jono_depth']}_{experiment_args['remove_gaussians']}_{experiment_args['sil_thres_gaussians']}_{experiment_args['l1_losses_embedding']}_{experiment_args['l1_losses_color']}_{experiment_args['bg_reg']}_{experiment_args['embeddings_lr']}_{experiment_args['red_lr']}_{experiment_args['red_lr_cam']}_{experiment_args['embedding_weight']}_{experiment_args['use_seg_for_nn']}_{experiment_args['weight_iso']}_{experiment_args['exp_weight']}_{experiment_args['loss_weight_emb']}_{experiment_args['loss_weight_iso']}_{experiment_args['loss_weight_rigid']}_{experiment_args['loss_weight_rot']}_{experiment_args['loss_weight_depth_cam']}_{experiment_args['forward_propagate_camera']}_{experiment_args['trafo_mat']}_{experiment_args['feats_224']}_{experiment_args['restart_if_fail']}_{experiment_args['early_stop']}_{experiment_args['stride']}_{experiment_args['time_window']}_{experiment_args['l1_losses_scale']}_{experiment_args['last_x']}_{experiment_args['kNN']}_{experiment_args['desired_image_height']}_{experiment_args['desired_image_width']}_{experiment_args['instseg_obj']}_{experiment_args['instseg_cam']}_{experiment_args['remove_close']}_transformed"
     else:
-        run_name = f"splatam_{seq}/splatam_{seq}_{experiment_args['seed']}_{experiment_args['mov_init_by']}_{experiment_args['tracking_iters']}_{experiment_args['tracking_iters_init']}_{experiment_args['tracking_iters_cam']}_{experiment_args['num_frames']}_{experiment_args['feature_dim']}_{experiment_args['remove_gaussians']}_{experiment_args['sil_thres_gaussians']}_{experiment_args['l1_losses_embedding']}_{experiment_args['l1_losses_color']}_{experiment_args['bg_reg']}_{experiment_args['embeddings_lr']}_{experiment_args['red_lr']}_{experiment_args['red_lr_cam']}_{experiment_args['embedding_weight']}_{experiment_args['use_seg_for_nn']}_{experiment_args['weight_iso']}_{experiment_args['exp_weight']}_{experiment_args['loss_weight_emb']}_{experiment_args['loss_weight_iso']}_{experiment_args['loss_weight_rigid']}_{experiment_args['loss_weight_rot']}_{experiment_args['loss_weight_depth_cam']}_{experiment_args['forward_propagate_camera']}_{experiment_args['trafo_mat']}_{experiment_args['feats_224']}_{experiment_args['restart_if_fail']}_{experiment_args['early_stop']}_{experiment_args['stride']}_{experiment_args['time_window']}_{experiment_args['l1_losses_scale']}_{experiment_args['last_x']}_{experiment_args['kNN']}_{experiment_args['desired_image_height']}_{experiment_args['desired_image_width']}_{experiment_args['instseg_obj']}_{experiment_args['instseg_cam']}_{experiment_args['smoothness']}_{experiment_args['prune_gaussians']}_{experiment_args['use_depth_error_for_adding_gaussians']}_{experiment_args['norm_embeddings']}_{experiment_args['add_only_bg']}_{experiment_args['remove_close']}_{experiment_args['loss_weight_depth_obj']}_aniso" #norm_embeddings
+        run_name = f"splatam_{seq}/splatam_{seq}_{experiment_args['seed']}_{experiment_args['mov_init_by']}_{experiment_args['tracking_iters']}_{experiment_args['tracking_iters_init']}_{experiment_args['tracking_iters_cam']}_{experiment_args['num_frames']}_{experiment_args['feature_dim']}_{experiment_args['remove_gaussians']}_{experiment_args['sil_thres_gaussians']}_{experiment_args['l1_losses_embedding']}_{experiment_args['l1_losses_color']}_{experiment_args['bg_reg']}_{experiment_args['embeddings_lr']}_{experiment_args['red_lr']}_{experiment_args['red_lr_cam']}_{experiment_args['embedding_weight']}_{experiment_args['use_seg_for_nn']}_{experiment_args['weight_iso']}_{experiment_args['exp_weight']}_{experiment_args['loss_weight_emb']}_{experiment_args['loss_weight_iso']}_{experiment_args['loss_weight_rigid']}_{experiment_args['loss_weight_rot']}_{experiment_args['loss_weight_depth_cam']}_{experiment_args['forward_propagate_camera']}_{experiment_args['trafo_mat']}_{experiment_args['feats_224']}_{experiment_args['restart_if_fail']}_{experiment_args['early_stop']}_{experiment_args['stride']}_{experiment_args['time_window']}_{experiment_args['l1_losses_scale']}_{experiment_args['last_x']}_{experiment_args['kNN']}_{experiment_args['desired_image_height']}_{experiment_args['desired_image_width']}_{experiment_args['instseg_obj']}_{experiment_args['instseg_cam']}_{experiment_args['smoothness']}_{experiment_args['prune_gaussians']}_{experiment_args['use_depth_error_for_adding_gaussians']}_{experiment_args['norm_embeddings']}_{experiment_args['add_only_bg']}_{experiment_args['remove_close']}_{experiment_args['loss_weight_depth_obj']}_{experiment_args['bg_loss']}_{aniso}_deb_l2_emb_r2" #norm_embeddings
 
     seq_experiment.config['run_name'] = run_name
     seq_experiment.config['data']['sequence'] = seq
@@ -228,7 +234,8 @@ def run_splatam(args):
     
     if experiment_args['bg_reg'] != 0:
         seq_experiment.config['tracking_obj']['loss_weights']['bg_reg'] = experiment_args['bg_reg']
-    
+    seq_experiment.config['tracking_obj']['loss_weights']['bg_loss'] = experiment_args['bg_loss']
+
     if experiment_args['embeddings_lr'] != 0:
         seq_experiment.config['tracking_obj']['lrs']['embeddings'] = experiment_args['embeddings_lr']
     if experiment_args['instseg_obj']:
@@ -260,6 +267,7 @@ def run_splatam(args):
     seq_experiment.config['tracking_cam']['forward_prop'] = experiment_args['forward_propagate_camera']
     seq_experiment.config['tracking_cam']['loss_weights']['depth'] = experiment_args['loss_weight_depth_cam']
     seq_experiment.config['tracking_cam']['loss_weights']['embeddings'] = experiment_args['loss_weight_emb']
+    seq_experiment.config['tracking_cam']['loss_weights']['im'] = experiment_args['loss_weight_img']
     seq_experiment.config['tracking_cam']['restart_if_fail'] = experiment_args['restart_if_fail']
     seq_experiment.config['exp_weight'] = experiment_args['exp_weight']
 
@@ -302,9 +310,9 @@ def run_splatam(args):
             return
         rgbd_slammer.eval()
     else:
-        # if os.path.isfile(os.path.join(results_dir, 'eval', 'traj_metrics.txt')):
-        #     print(f"Experiment already done {run_name}\n\n")
-        #     return
+        if os.path.isfile(os.path.join(results_dir, 'eval', 'traj_metrics.txt')):
+            print(f"Experiment already done {run_name}\n\n")
+            return
         
         os.makedirs(results_dir, exist_ok=True)
         shutil.copy(config_file, os.path.join(results_dir, "config.py"))
@@ -333,31 +341,32 @@ if __name__ == "__main__":
         num_frames = experiment.config['data']['num_frames'],
         dyno_losses = experiment.config['tracking_obj']['dyno_losses'],
         just_eval = experiment.config['just_eval'],
-        vis_all = True,
+        vis_all = False,
         vis_gt = False,
-        tracking_iters = 500,
-        tracking_iters_init = 1000,
-        tracking_iters_cam = 0,
+        tracking_iters = 200,
+        tracking_iters_init = 200,
+        tracking_iters_cam = 200,
         refine_iters = 0,
         mag_iso = True,
-        init_jono = True,
-        jono_depth = True,
-        l1_losses_embedding = 20,
-        l1_losses_color = 20, # 0.01,
+        init_jono = False,
+        jono_depth = False,
+        l1_losses_embedding = 0,
+        l1_losses_color = 0, # 0.01,
         bg_reg = 5,
         embeddings_lr = 0.001,
-        red_lr = False,
-        red_lr_cam = False,
+        red_lr = True,
+        red_lr_cam = True,
         remove_gaussians = False,
         sil_thres_gaussians = 0.5,
         make_grad_bg_smaller = 0,
         remove_outliers_l2 = 100,
-        embedding_weight = False,
+        embedding_weight = True,
         use_seg_for_nn = True,
         weight_iso = True,
         exp_weight = 2000,
         loss_weight_iso = 16,
         loss_weight_emb = 16,
+        loss_weight_img = 1,
         loss_weight_rigid = 128,
         loss_weight_rot = 16,
         loss_weight_depth_cam=0.1,
@@ -375,8 +384,8 @@ if __name__ == "__main__":
         base_transformations=False,
         base_transformations_mlp=False,
         kNN=20,
-        desired_image_height=360, # 240, # 120, #240, #480, 180, 360
-        desired_image_width=640, # 455, # 227, # 455, #910, 320, 640
+        desired_image_height=240, # 240, # 120, #240, #480, 180, 360
+        desired_image_width=455, # 455, # 227, # 455, #910, 320, 640
         instseg_obj=0.0,
         instseg_cam=0.0,
         smoothness=0,
@@ -384,8 +393,10 @@ if __name__ == "__main__":
         use_depth_error_for_adding_gaussians=False,
         norm_embeddings=True,
         add_only_bg=False,
-        remove_close=False,
-        loss_weight_depth_obj=0.1
+        remove_close=True,
+        loss_weight_depth_obj=0.1,
+        bg_loss=3,
+        aniso=True
         )
     
     davis_seqs = [
@@ -420,20 +431,22 @@ if __name__ == "__main__":
         'loading',
         'horsejump-high']
     
-    # davis_seqs = ['india']
+    # davis_seqs = ['motocross-jump']
     
     jono_seqs = ["boxes/ims/27", "softball/ims/27", "basketball/ims/21", "football/ims/18", "juggle/ims/14", "tennis/ims/8"]
 
+    rgb = [0] # list(range(50))
+
     configs_to_paralellize = list()
-    for seq in jono_seqs:
+    for seq in davis_seqs:
         # copy config and get create runname
         configs_to_paralellize.append([args.experiment, seq, experiment_args])
     
     # n_ranks = min(torch.cuda.device_count(), len(configs_to_paralellize))
     # gpus = ','.join([str(i) for i in range(n_ranks)])
     
-    n_ranks = 2
-    gpus = [4,6]
+    n_ranks = 6
+    gpus = [2,3,4,5,6,7]
 
     gpu_map(
         run_splatam,
